Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
beautifulsoup4==4.12.3
sqlalchemy==2.0.25
supabase>=2.9.0
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx<0.26,>=0.24
lxml==5.1.0
